Add foundation for rspec suite
To test an Engine, a dummy app has to be created.
This commit adds a development dependency on the Combustion gem that helps to bootstrap that dummy app.

It also adds rspec and ruby-debug19 as development dependencies.

The dummy app defines two models, User and Post- this is so that rails-erd doesn't fail to generate a diagram in the test environment.